Add files via upload
I missed the possibility to create circular references in FPGAs, for example for combining two NOR-gates to an SR-latch. Discrete implementation of latches with mesecon gates or implementations, that split latches and controlling logic into two or more FPGAs can lead to a situation, where more that one on/off-status travel through the circle creating a "flickering" feedback swinging. To avoid the problem when designing with discrete gates I made an own mod, but most of the logic in my circuits is packed in FPGAs for speed reasons, so I had to modify the FPGA code to enable this.
The regs are now persistent via the meta object. The function lg.interpret() expects them as a parameter and the calling function plg.update() creates them by getting the meta string with the name "regss" (registers string), which consists of 10 '1' or '0' characters representing the 10 numbered registers in the FPGA. 
With this alone, the FPGA could act as a shift-register, when the registers were just aligned in a series of register assignments, but this should not be the behaviour. To make complete changes when using forward-references with registers, the loop that interprets t, is nested into a loop that repeats, until no register change happens (or a safety counter is exceeded). If the safety counter gets exceeded, this gets logged in debug.txt. This is why the function lg.interpret() has gotten another additional parameter, the position, to do proper reporting.
After return, the registers, that were passed by reference, are taken to build a new regss-string, that is stored in meta.
First tests went well, but for experience I declare this as under construction.
I also made some minor changes to access of the operations table. I introduced 3 maps that contain indexes for operations, to make use of luas associative arrays and avoid loops that search for table-entries and 3 functions, that use that map an return indexes.