Add files via upload
Added a 10th register with the the special name "L" (for LED) instead of a number. In this register gets set, another texture is chosen and the FPGA illuminates. For this, another set of node-types for the FPGA get registered, the same as before plus the appendix "_led_on" for each of them.