Add files via upload
Added two fields for FPGA and FPGA-programmer: name and description. These can be altered in the FPGA-Dialog, they are copied with the FPGA-programmer and the field name is used for item-infotext and item-description. The information ist stored in the programmer's imeta and the FPGAs meta indexed by "name_s" and "desc_s".